+++ best_theme/example/lib/my_theme.g.dart
@@ -16,7 +16,14 @@ part of 'my_theme.dart';
 
 class _$Mytheme {
   late List<BestColor> myColors;
+  final Map<String, Map<String, Color>> _colors = {};
   bool isDark = false;
+  late Color color1;
+  late Color color2;
+  late Color color3;
+  late Color color4;
+  late Color grey23;
+  late Color grey3;
   List<ThemeMode> typeOfThemes = [
     ThemeMode.system,
     ThemeMode.light,
@@ -28,6 +35,17 @@ class _$Mytheme {
   late final ValueNotifier<ThemeParam> _themeNotifier;
 
   _$Mytheme({List<BestColor>? myColors, ThemeMode mode = ThemeMode.system}) {
+    if (myColors != null) {
+      for (var color in myColors) {
+        _colors[color.name] = color.toMap();
+      }
+    }
+    color1 = _colors['color1']![isDark ? 'dark' : 'light']!;
+    color2 = _colors['color2']![isDark ? 'dark' : 'light']!;
+    color3 = _colors['color3']![isDark ? 'dark' : 'light']!;
+    color4 = _colors['color4']![isDark ? 'dark' : 'light']!;
+    grey23 = _colors['grey23']![isDark ? 'dark' : 'light']!;
+    grey3 = _colors['grey3']![isDark ? 'dark' : 'light']!;
     if (myColors != null) {
       this.myColors = myColors;
     }
@@ -132,6 +150,7 @@ class _$Mytheme {
 
   void toggle() {
     isDark = !isDark;
+    toggleColor();
     _themeNotifier.value =
         ThemeParam(isDark ? ThemeMode.dark : ThemeMode.light, myColors, isDark);
     currentTheme = _themeNotifier.value.mode;
@@ -139,18 +158,21 @@ class _$Mytheme {
 
   void toDark() {
     isDark = true;
+    toggleColor();
     _themeNotifier.value = ThemeParam(ThemeMode.dark, myColors, true);
     currentTheme = _themeNotifier.value.mode;
   }
 
   void toLight() {
     isDark = false;
+    toggleColor();
     _themeNotifier.value = ThemeParam(ThemeMode.light, myColors, false);
     currentTheme = _themeNotifier.value.mode;
   }
 
   void toSystem(BuildContext context) {
     isDark = MediaQuery.of(context).platformBrightness == Brightness.dark;
+    toggleColor();
     _themeNotifier.value = ThemeParam(ThemeMode.system, myColors, isDark);
     currentTheme = _themeNotifier.value.mode;
   }
@@ -162,6 +184,7 @@ class _$Mytheme {
     } else {
       isDark = mode == ThemeMode.dark;
     }
+    toggleColor();
     _themeNotifier.value = ThemeParam(mode, myColors, isDark);
     currentTheme = _themeNotifier.value.mode;
   }
@@ -176,6 +199,15 @@ class _$Mytheme {
   Color primaryColorDark(context) => Theme.of(context).primaryColorDark;
   Color primaryColorLight(context) => Theme.of(context).primaryColorLight;
   ThemeData theme(context) => Theme.of(context);
+
+  toggleColor() {
+    color1 = _colors['color1']![isDark ? 'dark' : 'light']!;
+    color2 = _colors['color2']![isDark ? 'dark' : 'light']!;
+    color3 = _colors['color3']![isDark ? 'dark' : 'light']!;
+    color4 = _colors['color4']![isDark ? 'dark' : 'light']!;
+    grey23 = _colors['grey23']![isDark ? 'dark' : 'light']!;
+    grey3 = _colors['grey3']![isDark ? 'dark' : 'light']!;
+  }
 
   Widget BestThemeBuilder({
     required Widget Function(ThemeParam theme, BuildContext context) builder,
@@ -194,7 +226,7 @@ class ThemeParam {
   late Color color2;
   late Color color3;
   late Color color4;
-  late Color grey2;
+  late Color grey23;
   late Color grey3;
   ThemeParam(this.mode, List<BestColor> myColors, bool isDark) {
     for (var color in myColors) {
@@ -204,7 +236,7 @@ class ThemeParam {
     color2 = _colors['color2']![isDark ? 'dark' : 'light']!;
     color3 = _colors['color3']![isDark ? 'dark' : 'light']!;
     color4 = _colors['color4']![isDark ? 'dark' : 'light']!;
-    grey2 = _colors['grey2']![isDark ? 'dark' : 'light']!;
+    grey23 = _colors['grey23']![isDark ? 'dark' : 'light']!;
     grey3 = _colors['grey3']![isDark ? 'dark' : 'light']!;
   }
 }