+++ best_theme/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2025-1-19
+### added
+- Added custom colors to main class
+- Support latest analyzer
+
 ## [0.0.2] - 2024-12-29
 ### Added
 - Support go router